requirements-prod.txt
anyascii==0.3.2
attrs==23.2.0
beautifulsoup4==4.11.2
CacheControl==0.14.0
certifi==2024.6.2
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.8
django-cors-headers==4.3.1
django-db-connection-pool==1.2.5
django-filter==23.5
django-modelcluster==6.3
django-permissionedforms==0.1
django-taggit==5.0.0
django-treebeard==4.7.1
djangorestframework==3.15.1
draftjs-exporter==2.1.7
drf-nested-routers==0.94.1
drf-spectacular==0.27.2
et-xmlfile==1.1.0
exceptiongroup==1.2.1
filelock==3.14.0
grimp==3.2
gunicorn==22.0.0
html5lib==1.1
identify==2.5.36
idna==3.7
importlib-metadata==7.1.0
inflection==0.5.1
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.4
jsonschema==4.22.0
kaleido==0.2.1
l18n==2021.3
lockfile==0.12.2
Markdown==3.6
markdown-it-py==2.2.0
MarkupSafe==2.1.5
mdurl==0.1.2
msgpack==1.0.8
mypy-extensions==1.0.0
nodeenv==1.9.1
numpy==1.26.4
openapi-codec==1.3.2
openpyxl==3.1.3
orjson==3.10.3
packageurl-python==0.15.0
packaging==24.0
pandas==2.2.1
pathspec==0.12.1
pbr==6.0.0
Pillow==10.3.0
platformdirs==4.2.2
plotly==5.22.0
pluggy==1.5.0
pyparsing==3.1.2
psycopg2==2.9.9
pyrsistent==0.20.0
python-dateutil==2.9.0.post0
pytz==2024.1
Pygments==2.18.0
PyYAML==6.0.1
redis==5.0.5
requests==2.32.3
rich==13.7.1
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
SQLAlchemy==2.0.30
scour==0.38.2
setuptools==70.0.0
simplejson==3.19.2
six==1.16.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlparams==6.0.1
telepath==0.3.1
tenacity==8.3.0
toml==0.10.2
tomli==2.0.1
typing_extensions==4.12.1
uritemplate==4.1.1
urllib3==2.0.7
virtualenv==20.26.2
wagtail==6.1.2
wagtail_trash==3.0.0
wagtail_modeladmin==2.0.0
webencodings==0.5.1
Willow==1.8.0
zipp==3.19.2
-r requirements-prod-ingestion.txt
